102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
  */
 int main(void)
 {
-long long int fib1 = 1, fib2 = 2, next;
+long int fib1 = 1, fib2 = 2, next;
 int count;
-printf("%lld, %lld", fib1, fib2);
+printf("%ld, %ld", fib1, fib2);
 for (count = 2; count < 50; count++)
 {
 next = fib1 + fib2;
-printf(", %lld", next);
+printf(", %ld", next);
 fib1 = fib2;
 fib2 = next;
 }
